# api/openapi.yaml
openapi: 3.1.0
info:
  title: Rackd API
  description: Infrastructure management API for devices, networks, datacenters, and discovery
  version: 1.0.0
  contact:
    name: Rackd
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        maximum: 1000
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: >
            One of: INVALID_JSON, INVALID_INPUT, VALIDATION_ERROR, NOT_FOUND,
            FORBIDDEN, UNAUTHORIZED, INVALID_CREDENTIALS, ALREADY_EXISTS,
            IP_NOT_AVAILABLE, CANNOT_DELETE_SELF, SYSTEM_ROLE,
            RATE_LIMIT_EXCEEDED, LOGIN_RATE_LIMIT_EXCEEDED, CSRF_FAILED,
            EXPIRED_KEY, LEGACY_API_KEY_UNSUPPORTED, INTERNAL_ERROR
        error:
          type: string
        details:
          type: object
          additionalProperties: true
    Datacenter:
      type: object
      required: [id, name, location, description, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        location: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      additionalProperties: true
    DatacenterInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        location: { type: string }
        description: { type: string }
...
    User:
      type: object
      required: [id, username, is_active, is_admin, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string }
        full_name: { type: string }
        is_active: { type: boolean }
        is_admin: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_login_at: { type: string, format: date-time }
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
      additionalProperties: true
...
    Role:
      type: object
      required: [id, name, is_system, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        is_system: { type: boolean }
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      additionalProperties: true
...
    Permission:
      type: object
      required: [id, name, resource, action, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        resource: { type: string }
        action: { type: string }
        created_at: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/datacenters:
    get:
      operationId: listDatacenters
      tags: [Datacenters]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: name
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of datacenters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Datacenter'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createDatacenter
      tags: [Datacenters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatacenterInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datacenter'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/datacenters/{id}:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getDatacenter
      tags: [Datacenters]
      responses:
        '200':
          description: Datacenter details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datacenter'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updateDatacenter
      tags: [Datacenters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatacenterInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datacenter'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteDatacenter
      tags: [Datacenters]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/auth/me:
    get:
      operationId: getCurrentUser
      tags: [Auth]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── Users ──
...